baneloforecasting/requirements.txt
# Django and database
Django>=4.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Firebase (if still used for mobile)
firebase-admin>=6.0.0

# ML/Data processing (optional, for forecasting)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
